separate new project for virtual server
