+++ src/CSharp/EasyMicroservices.SMS.sln
@@ -17,6 +17,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EasyMicroservices.SMS.Tests
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EasyMicroservices.SMS.Kavenegar", "EasyMicroservices.SMS.Kavenegar\EasyMicroservices.SMS.Kavenegar.csproj", "{6E98827F-6165-4F82-B7F9-E3DB09D60D85}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EasyMicroservices.SMS.VirtualServerForTests", "EasyMicroservices.SMS.VirtualServerForTests\EasyMicroservices.SMS.VirtualServerForTests.csproj", "{7A34CB55-FFE8-4D68-A233-E16C39E91CFE}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -63,6 +65,18 @@ Global
 		{6E98827F-6165-4F82-B7F9-E3DB09D60D85}.Release|x64.Build.0 = Release|Any CPU
 		{6E98827F-6165-4F82-B7F9-E3DB09D60D85}.Release|x86.ActiveCfg = Release|Any CPU
 		{6E98827F-6165-4F82-B7F9-E3DB09D60D85}.Release|x86.Build.0 = Release|Any CPU
+		{7A34CB55-FFE8-4D68-A233-E16C39E91CFE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7A34CB55-FFE8-4D68-A233-E16C39E91CFE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7A34CB55-FFE8-4D68-A233-E16C39E91CFE}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{7A34CB55-FFE8-4D68-A233-E16C39E91CFE}.Debug|x64.Build.0 = Debug|Any CPU
+		{7A34CB55-FFE8-4D68-A233-E16C39E91CFE}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{7A34CB55-FFE8-4D68-A233-E16C39E91CFE}.Debug|x86.Build.0 = Debug|Any CPU
+		{7A34CB55-FFE8-4D68-A233-E16C39E91CFE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7A34CB55-FFE8-4D68-A233-E16C39E91CFE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7A34CB55-FFE8-4D68-A233-E16C39E91CFE}.Release|x64.ActiveCfg = Release|Any CPU
+		{7A34CB55-FFE8-4D68-A233-E16C39E91CFE}.Release|x64.Build.0 = Release|Any CPU
+		{7A34CB55-FFE8-4D68-A233-E16C39E91CFE}.Release|x86.ActiveCfg = Release|Any CPU
+		{7A34CB55-FFE8-4D68-A233-E16C39E91CFE}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -74,6 +88,7 @@ Global
 		{D3A32316-468F-49C5-9AD1-394071FACC7A} = {988923C0-A6DE-4F3B-BBF5-21B1DC9BB98D}
 		{AE1E0363-39A7-4B72-972A-12E8F33F74F4} = {D3A32316-468F-49C5-9AD1-394071FACC7A}
 		{6E98827F-6165-4F82-B7F9-E3DB09D60D85} = {37E2C62E-365E-4D2C-8E64-9456132D4048}
+		{7A34CB55-FFE8-4D68-A233-E16C39E91CFE} = {D3A32316-468F-49C5-9AD1-394071FACC7A}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {2973C221-95FB-4EE5-A108-9CB8FB1E4698}
